+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Support for `dbVersion` field in Autonomous databases Container database resources
 - Support for Archive Log Backup and Point in time restore
 - Support resource discovery for `datacatalog` resources
+- Support resource discovery for `integration` resources
+
 
 ### Fixed
 - Fixed plan failure in case of missing required attributes in resource discovery. Placeholder values will be added for missing required attributes and the attributes will be added to `lifecycle ignore_changes`